rbd: store snapshot id instead of index into changing array
When a device was open at a snapshot, and snapshots were deleted or
added, data from the wrong snapshot would be read. Instead of assuming
the snap context is constant, store the actual snap id when the device
is initialized, and rely on the OSDs to signal an error if we try
reading from a snapshot that was deleted.

Signed-off-by: Josh Durgin <josh.durgin@dreamhost.com>