libceph: define ceph_decode_string()
There is no string decoding function defined in <decode.h>.  This
function is a little different from the others in that the length
of the encoded string is not known a priori.  So the interface is
defined a bit like snprintf(), where the value returned indicates
the space required--even if it's more than the space allotted.

Signed-off-by: Alex Elder <elder@inktank.com>